% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.R
\name{top}
\alias{top}
\title{Print the top of a matrix}
\usage{
top(x, n)
}
\arguments{
\item{x}{a \code{matrix}}

\item{n}{Number of rows and columns to display of the matrix}
}
\value{
A \code{NULL}
}
\description{
\code{top} is a small function to not just present the first rows
}
\examples{
mat = matrix(rnorm(100), nrow = 10)
top(mat, n = 4)
\dontrun{
   top(mat, n = 10)
}
}
